base-files: only rm /home/root if it exists
the debian package do not have a /home/root, and hence this fails

+++ meta-venus/recipes-core/base-files/base-files_%.bbappend
@@ -7,6 +7,8 @@ dirs755 += " /data"
 
 # Replace home dir with symlink to persistent volume
 do_install_append() {
-    rmdir ${D}/home/root
-    ln -s ${permanentdir}/home/root ${D}/home/root
+	if [ -d ${D}/home/root ]; then
+		rmdir ${D}/home/root
+		ln -s ${permanentdir}/home/root ${D}/home/root
+	fi
 }
